requirements.txt
graphql-core==3.2.0
